# wip: perf: many many small fixes