--- a/frappe/database/database.py
+++ b/frappe/database/database.py
@@ -339,9 +339,9 @@ class Database:
 		"""Takes the query and logs it to various interfaces according to the settings."""
 		_query = None
 
-		if frappe.conf.allow_tests and frappe.cache.get_value("flag_print_sql"):
-			_query = _query or str(mogrified_query)
-			print(_query)
+		# if frappe.conf.allow_tests and frappe.cache.get_value("flag_print_sql"):
+		# 	_query = _query or str(mogrified_query)
+		# 	print(_query)
 
 		if debug:
 			_query = _query or str(mogrified_query)