--- a/frappe/database/mariadb/database.py
+++ b/frappe/database/mariadb/database.py
@@ -1,5 +1,5 @@
-import re
 import datetime
+import re
 from contextlib import contextmanager
 
 import pymysql