--- a/frappe/installer.py
+++ b/frappe/installer.py
@@ -601,6 +601,7 @@ def make_site_config(
 def update_site_config(key, value, validate=True, site_config_path=None):
 	"""Update a value in site_config"""
 	from frappe.utils.synchronization import filelock
+	from frappe import _get_site_config, get_common_site_config
 
 	if not site_config_path:
 		site_config_path = get_site_config_path()
@@ -610,6 +611,9 @@ def update_site_config(key, value, validate=True, site_config_path=None):
 
 	with filelock("site_config", is_global=_is_global_conf):
 		_update_config_file(key=key, value=value, config_file=site_config_path)
+
+	_get_site_config.clear_cache()
+	get_common_site_config.clear_cache()
 
 
 def _update_config_file(key: str, value, config_file: str):