This PR is not meant to be merged like this. I needed a set of correct-ish changes that are wide enough to remove overheads that *any kind* of test workload will show significant delta improvements. If you want to understand changes then go commit by commit.

Each commit is likely worthy of a separate PR w/ test and rigorous review. So for now, just ignore this. 

Some reference data of "ping" endpoint (I am treating this as "overhead")

| Metric     | Baseline | This PR |  Speedup |
|  ---       |    ---      | ---        |   ---     |
| Throughput (rps) |  261        |  318       | 1.21         |
| Latency (ms)    | 224         | 184       | 1.21          |

Baseline for all future changes: 7c06c845ea345a67fec66d0321dedf9a1bbc2152 

My hunch is 21% difference here is enough to show up as >1% in all kinds of workloads. So this is good enough to write benchmark suite and do A-A, A-B comparisons.

Test setup:
- Single gunicorn worker pinned on one core: `taskset -c 0 gunicorn -b 127.0.0.1:8000 -w 1 --chdir /home/ankush/benches/develop/sites 'frappe.app:application_with_statics()' --reuse-port`
- (mostly) Quiesced local setup w/ <1% variance in measurements - https://ankush.dev/p/reliable-benchmarking
- load generator - `wrk -t6 -c60 -d5 http://dev.localhost:8000/api/method/ping`