@ankush 2024-12-13
```
bench_database_empty_transaction_cycling: Mean +- std dev: [before] 286 us +- 4 us -> [after] 280 us +- 3 us: 1.02x faster
bench_database_get_single_value: Mean +- std dev: [before] 1.24 ms +- 0.01 ms -> [after] 829 us +- 9 us: 1.50x faster
bench_database_get_value_simple: Mean +- std dev: [before] 5.64 ms +- 0.06 ms -> [after] 3.62 ms +- 0.04 ms: 1.56x faster
bench_database_select_star: Mean +- std dev: [before] 2.51 ms +- 0.03 ms -> [after] 1.59 ms +- 0.02 ms: 1.57x faster
bench_orm_get_doc: Mean +- std dev: [before] 10.8 ms +- 0.2 ms -> [after] 8.64 ms +- 0.12 ms: 1.25x faster
bench_orm_get_local_cached_doc: Mean +- std dev: [before] 29.7 us +- 0.1 us -> [after] 21.6 us +- 0.3 us: 1.38x faster
bench_orm_get_user: Mean +- std dev: [before] 16.3 ms +- 0.3 ms -> [after] 12.1 ms +- 0.1 ms: 1.34x faster
bench_redis_make_key: Mean +- std dev: [before] 152 us +- 1 us -> [after] 69.8 us +- 0.6 us: 2.18x faster

Benchmark hidden because not significant (3): bench_database_get_cached_value_simple, bench_orm_get_cached_doc, bench_redis_set_value

Geometric mean: 1.31x faster
```

Comparison of few microbench from https://github.com/ankush/caffeination